requirements.txt
Flask
Flask[async]
Flask-Login
Flask-SQLAlchemy
python-dotenv
httpx
Werkzeug